+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Added a button in the settings to automatically set correct options to keep compatibility with Dendron.
 - Added confirmation dialog when resetting properties or setting Dendron compatibility options.
 - A new command for creating a note will open lookup with the currently open note name auto-filled.
+- Fixed a bug that could cause a crash if title was number type.
 
 ## 1.5.2